requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
pymysql==1.0.2
PyJWT==2.8.0
DBUtils==3.0.0
Werkzeug==2.0.1
python-dotenv==1.0.0
python-docx==0.8.11
reportlab==4.0.7
Pillow==10.0.0
